js: Consolidate exception definitions into exceptions.ts
It's silly to have them split across multiple files
