--- js/src/exceptions/not_implemented_error.ts
@@ -1,7 +1,0 @@
-/** Thrown when we attempt to use an unsupported crypto algorithm via WebCrypto */
-export class NotImplementedError extends Error {
-  constructor(m: string) {
-    super(m);
-    Object.setPrototypeOf(this, NotImplementedError.prototype);
-  }
-}
